+++ common/src/main/java/traben/entity_model_features/mixin/MixinBlockEntityRendererFactories.java
@@ -9,6 +9,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import traben.entity_model_features.EMFClient;
 import traben.entity_model_features.utils.EMFManager;
 
 @Mixin(BlockEntityRendererFactories.class)
@@ -17,6 +18,8 @@ public class MixinBlockEntityRendererFactories {
     @Inject(method = "method_32145", at = @At(value = "HEAD"))
     private static void setEmf$Model(ImmutableMap.Builder<?, ?> builder, BlockEntityRendererFactory.Context context, BlockEntityType<?> type, BlockEntityRendererFactory<?> factory, CallbackInfo ci) {
         //mark which variant is currently specified for use by otherwise identical block entity renderers
+        if (EMFClient.testForForgeLoadingError()) return;
+
         if (BlockEntityType.CHEST.equals(type))
             EMFManager.getInstance().currentSpecifiedModelLoading = "chest";
         else if (BlockEntityType.ENDER_CHEST.equals(type))