+++ common/src/main/java/traben/entity_model_features/mixin/rendering/MixinWolfCollarFeatureRenderer.java
@@ -14,11 +14,10 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import traben.entity_model_features.EMFVersionDifferenceManager;
+import traben.entity_model_features.EMFClient;
 import traben.entity_model_features.models.EMFModelPartRoot;
 import traben.entity_model_features.models.IEMFModel;
 import traben.entity_model_features.utils.EMFManager;
-import traben.entity_model_features.utils.EMFUtils;
 
 @Mixin(WolfCollarFeatureRenderer.class)
 public class MixinWolfCollarFeatureRenderer {
@@ -29,23 +28,13 @@ public class MixinWolfCollarFeatureRenderer {
 
     @Inject(method = "<init>", at = @At("TAIL"))
     private void setEmf$Model(FeatureRendererContext<?, ?> featureRendererContext, CallbackInfo ci) {
-        ModelPart w;
-        if (EMFVersionDifferenceManager.isForge()) {
-            try {
-                w = EMFManager.getInstance().injectIntoModelRootGetter(emf$collar_layer, WolfEntityModel.getTexturedModelData().createModel());
-            } catch (IncompatibleClassChangeError error) {
-                EMFUtils.logError("///////////////////");
-                EMFUtils.logError("EMF crashed due to a forge dependency error (probably), suppressing the EMF crash so the true culprit will be sent to the crash report tool\nIF THIS HAPPENS MORE THAN ONCE THIS MIGHT ACTUALLY BE AN EMF ISSUE\n");
-                error.printStackTrace();
-                EMFUtils.logError("\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\");
-                return;
-            }
-        } else {
-            w = EMFManager.getInstance().injectIntoModelRootGetter(emf$collar_layer, WolfEntityModel.getTexturedModelData().createModel());
-        }
+        if(EMFClient.testForForgeLoadingError()) return;
+
+        ModelPart collarModel = EMFManager.getInstance().injectIntoModelRootGetter(emf$collar_layer, WolfEntityModel.getTexturedModelData().createModel());
+
         //separate the collar model, if it has a custom jem model or the base wolf has a custom jem model
-        if (w instanceof EMFModelPartRoot || ((IEMFModel) featureRendererContext.getModel()).emf$isEMFModel()) {
-            EMFManager.wolfCollarModel = new WolfEntityModel<>(w);
+        if (collarModel instanceof EMFModelPartRoot || ((IEMFModel) featureRendererContext.getModel()).emf$isEMFModel()) {
+            EMFManager.wolfCollarModel = new WolfEntityModel<>(collarModel);
         }
     }