+++ common/src/main/java/traben/entity_model_features/models/EMFModelPart.java
@@ -8,16 +8,14 @@ import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.WorldRenderer;
-import net.minecraft.client.render.item.ItemRenderer;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Box;
 import traben.entity_model_features.models.animation.EMFAnimationHelper;
 import traben.entity_model_features.utils.EMFManager;
-import traben.entity_texture_features.ETFClientCommon;
-import traben.entity_texture_features.features.ETFManager;
 import traben.entity_texture_features.features.ETFRenderContext;
 import traben.entity_texture_features.features.texture_handlers.ETFTexture;
+import traben.entity_texture_features.utils.ETFUtils2;
 import traben.entity_texture_features.utils.ETFVertexConsumer;
 
 import java.util.HashMap;
@@ -74,9 +72,9 @@ public abstract class EMFModelPart extends ModelPart {
             if(newConsumer instanceof ETFVertexConsumer newETFConsumer){
                 ETFTexture etfTexture = newETFConsumer.etf$getETFTexture();
                 if(etfTexture == null) return;
-
-                etf$renderEmissive(etfTexture, provider, matrices, overlay, red, green, blue, alpha);
-                etf$renderEnchanted(etfTexture, provider, matrices, light, overlay, red, green, blue, alpha);
+                ETFUtils2.RenderMethodForOverlay renderMethodForOverlay = (prov,ligh)-> renderLikeVanilla(matrices,prov,ligh,overlay,red,green,blue,alpha);
+                ETFUtils2.renderEmissive(etfTexture, provider, renderMethodForOverlay);
+                ETFUtils2.renderEnchanted(etfTexture, provider, light, renderMethodForOverlay);
             }
 
 
@@ -130,8 +128,9 @@ public abstract class EMFModelPart extends ModelPart {
                     //are these render required objects valid?
                     if (provider != null && layer != null) {
                         //attempt special renders as eager OR checks
-                        if (etf$renderEmissive(texture, provider, matrices, overlay, red, green, blue, alpha) |
-                                etf$renderEnchanted(texture, provider, matrices, light, overlay, red, green, blue, alpha)) {
+                        ETFUtils2.RenderMethodForOverlay renderMethodForOverlay = (prov,ligh)-> renderLikeVanilla(matrices,prov,ligh,overlay,red,green,blue,alpha);
+                        if (ETFUtils2.renderEmissive(texture, provider, renderMethodForOverlay) |
+                                ETFUtils2.renderEnchanted(texture, provider, light, renderMethodForOverlay)) {
                             //reset render layer stuff behind the scenes if special renders occurred
                             //this will also return ETFVertexConsumer held data to normal if the same ETFVertexConsumer
                             //was previously affected by a special render
@@ -284,51 +283,6 @@ public abstract class EMFModelPart extends ModelPart {
         return mapOfAll;
     }
 
-    //copy of etf rewrite emissive rendering code
-
-    private boolean etf$renderEmissive(ETFTexture texture, VertexConsumerProvider provider, MatrixStack matrices, int overlay, float red, float green, float blue, float alpha) {
-        Identifier emissive = texture.getEmissiveIdentifierOfCurrentState();
-        if (emissive != null) {
-            boolean wasAllowed = ETFRenderContext.isAllowedToRenderLayerTextureModify();
-            ETFRenderContext.preventRenderLayerTextureModify();
-
-            boolean textureIsAllowedBrightRender = ETFManager.getEmissiveMode() == ETFManager.EmissiveRenderModes.BRIGHT
-                    && ETFRenderContext.getCurrentEntity().etf$canBeBright();// && !ETFRenderContext.getCurrentETFTexture().isPatched_CurrentlyOnlyArmor();
-
-            VertexConsumer emissiveConsumer = provider.getBuffer(
-                    textureIsAllowedBrightRender ?
-                            RenderLayer.getBeaconBeam(emissive, true) :
-                            ETFRenderContext.getCurrentEntity().etf$isBlockEntity() ?
-                                    RenderLayer.getEntityTranslucentCull(emissive) :
-                                    RenderLayer.getEntityTranslucent(emissive));
-
-            if(wasAllowed) ETFRenderContext.allowRenderLayerTextureModify();
-
-            ETFRenderContext.startSpecialRenderOverlayPhase();
-            renderLikeVanilla(matrices, emissiveConsumer, ETFClientCommon.EMISSIVE_FEATURE_LIGHT_VALUE, overlay, red, green, blue, alpha);
-            ETFRenderContext.endSpecialRenderOverlayPhase();
-            return true;
-        }
-        return false;
-    }
-
-
-    private boolean etf$renderEnchanted(ETFTexture texture, VertexConsumerProvider provider, MatrixStack matrices, int light, int overlay, float red, float green, float blue, float alpha) {
-        //attempt enchanted render
-        Identifier enchanted = texture.getEnchantIdentifierOfCurrentState();
-        if (enchanted != null) {
-            boolean wasAllowed = ETFRenderContext.isAllowedToRenderLayerTextureModify();
-            ETFRenderContext.preventRenderLayerTextureModify();
-            VertexConsumer enchantedVertex = ItemRenderer.getArmorGlintConsumer(provider, RenderLayer.getArmorCutoutNoCull(enchanted), false, true);
-            if(wasAllowed) ETFRenderContext.allowRenderLayerTextureModify();
-
-            ETFRenderContext.startSpecialRenderOverlayPhase();
-            renderLikeVanilla(matrices, enchantedVertex, light, overlay, red, green, blue, alpha);
-            ETFRenderContext.endSpecialRenderOverlayPhase();
-            return true;
-        }
-        return false;
-    }
 
 
     public static class Animator implements Runnable {